Restored full functionality on NVIDIA
